Hex27 SIMD MasterElements
* Miscellaneous cleanups to the master element section
