+++ include/BuildTemplates.h
@@ -11,6 +11,11 @@
 
 #include <AlgTraits.h>
 
+#ifndef USER_POLY_ORDER
+#define USER_POLY_ORDER 5
+#endif
+
+
 namespace sierra{
 namespace nalu{
 
@@ -26,17 +31,17 @@ template class ClassName<AlgTraitsQuad4_2D>;                      \
 template class ClassName<AlgTraitsQuad9_2D>;                      \
 template class ClassName<AlgTraitsTri3_2D>;                       \
 
-#define INSTANTIATE_KERNEL_3D_HO(ClassName)                       \
-template class ClassName<AlgTraitsHexGL<2>>;                      \
-template class ClassName<AlgTraitsHexGL<3>>;                      \
-template class ClassName<AlgTraitsHexGL<4>>;                      \
-template class ClassName<AlgTraitsHexGL<5>>;                      \
+#define INSTANTIATE_KERNEL_3D_HO(ClassName)                      \
+template class ClassName<AlgTraitsHexGL<2>>;                     \
+template class ClassName<AlgTraitsHexGL<3>>;                     \
+template class ClassName<AlgTraitsHexGL<4>>;                     \
+template class ClassName<AlgTraitsHexGL<USER_POLY_ORDER>>;       \
 
-#define INSTANTIATE_KERNEL_2D_HO(ClassName)                       \
-template class ClassName<AlgTraitsQuadGL<2>>;                     \
-template class ClassName<AlgTraitsQuadGL<3>>;                     \
-template class ClassName<AlgTraitsQuadGL<4>>;                     \
-template class ClassName<AlgTraitsQuadGL<5>>;                     \
+#define INSTANTIATE_KERNEL_2D_HO(ClassName)                      \
+template class ClassName<AlgTraitsQuadGL<2>>;                    \
+template class ClassName<AlgTraitsQuadGL<3>>;                    \
+template class ClassName<AlgTraitsQuadGL<4>>;                    \
+template class ClassName<AlgTraitsQuadGL<USER_POLY_ORDER>>;      \
 
 #define INSTANTIATE_KERNEL(ClassName)           \
   INSTANTIATE_KERNEL_3D(ClassName)              \
