+++ reg_tests/CTestList.cmake
@@ -87,7 +87,7 @@ add_test_r(hoHelium 8)
 add_test_r(hoVortex 2)
 add_test_r(inputFireEdgeUpwind 4)
 add_test_r(inputFireElem 4)
-add_test_r(kovasznay_P7 1)
+add_test_r(kovasznay_P5 1)
 add_test_r(milestoneRun 4)
 add_test_r(milestoneRunConsolidated 4)
 add_test_r(mixedTetPipe 8)
