+++ reg_tests/test_files/hoVortex_P2/hoVortex_P2_R2.i
@@ -120,10 +120,8 @@ realms:
     solution_options:
       name: myOptions
       turbulence_model: laminar
-      use_consolidated_solver_algorithm: yes
-  
-      options:
 
+      options:
         - hybrid_factor:
             velocity: 0.0
 
@@ -132,7 +130,7 @@ realms:
             velocity: no
 
         - element_source_terms:
-            momentum: [momentum_time_derivative, advection_diffusion, SteadyTaylorVortex]
+            momentum: momentum_time_derivative
 
         - consistent_mass_matrix_png:
             pressure: yes
