+++ reg_tests/test_files/kovasznay_P5/kovasznay_P5.norm.gold
@@ -1,0 +1,100 @@
+0.0106973360889386 1 0.001
+0.01118572962432006 2 0.002
+0.002931128793911018 3 0.003
+0.000652804306730063 4 0.004
+0.0005198118890440314 5 0.005
+0.0005259465488241787 6 0.006
+0.0005385294419189139 7 0.007
+0.0004649289430840002 8 0.008
+0.0003174211055418049 9 0.009
+0.0001740002160560426 10 0.01
+0.0001888213694809032 11 0.011
+0.0002717282683369481 12 0.012
+0.0003055379059200231 13 0.013
+0.0002811131299588282 14 0.014
+0.0002168047105235382 15 0.015
+0.0001424769039882773 16 0.016
+0.0001104462699107147 17 0.017
+0.0001378902389432126 18 0.018
+0.0001619042514745417 19 0.019
+0.000159512537945568 20 0.02
+0.0001349605244218945 21 0.021
+0.0001026592091286429 22 0.022
+8.296782140405219e-05 23 0.023
+8.556528376894573e-05 24 0.024
+9.294462483104789e-05 25 0.025
+9.137783153206565e-05 26 0.026
+8.038443469218413e-05 27 0.027
+6.651147733460453e-05 28 0.028
+5.833412784103971e-05 29 0.029
+5.785750034756074e-05 30 0.03
+5.820720707827357e-05 31 0.031
+5.429187783313012e-05 32 0.032
+4.671864633201715e-05 33 0.033
+4.011949551591842e-05 34 0.034
+3.893130097069111e-05 35 0.035
+4.09997802259515e-05 36 0.036
+4.114373233722504e-05 37 0.037
+3.723655515097819e-05 38 0.038
+3.050460569591427e-05 39 0.039
+2.497216115717438e-05 40 0.04
+2.498292949378565e-05 41 0.041
+2.79448867276626e-05 42 0.042
+2.903736954663112e-05 43 0.043
+2.67130303300633e-05 44 0.044
+2.182166275614514e-05 45 0.045
+1.720741014091118e-05 46 0.046
+1.706336406116275e-05 47 0.047
+1.988004916951965e-05 48 0.048
+2.146495428835815e-05 49 0.049
+2.052110663080863e-05 50 0.05
+1.746226871547363e-05 51 0.051
+1.388678514392016e-05 52 0.052
+1.270960817070304e-05 53 0.053
+1.439414692375843e-05 54 0.054
+1.584744468594011e-05 55 0.055
+1.57638209524994e-05 56 0.056
+1.425089053657401e-05 57 0.057
+1.223581820638917e-05 58 0.058
+1.124986966751312e-05 59 0.059
+1.193861424735729e-05 60 0.06
+1.280665785374205e-05 61 0.061
+1.288324574048467e-05 62 0.062
+1.207174178408264e-05 63 0.063
+1.083573103725066e-05 64 0.064
+9.984931160621706e-06 65 0.065
+1.009152533013625e-05 66 0.066
+1.055683580282932e-05 67 0.067
+1.072552131445821e-05 68 0.068
+1.042988134155499e-05 69 0.069
+9.877132105528992e-06 70 0.07
+9.443290123757227e-06 71 0.071
+9.385321819947939e-06 72 0.072
+9.521508185673438e-06 73 0.073
+9.540300066909117e-06 74 0.074
+9.323989677607705e-06 75 0.075
+8.955593407879206e-06 76 0.076
+8.626355000845957e-06 77 0.077
+8.497869837373721e-06 78 0.078
+8.520891493218612e-06 79 0.079
+8.534192028803799e-06 80 0.08
+8.452527178006305e-06 81 0.081
+8.292817998578286e-06 82 0.082
+8.131104349620801e-06 83 0.083
+8.03184497506154e-06 84 0.084
+7.988261436966655e-06 85 0.085
+7.934707369577911e-06 86 0.086
+7.831707138903323e-06 87 0.087
+7.687541406842739e-06 88 0.088
+7.542937412623941e-06 89 0.089
+7.438939326055962e-06 90 0.09
+7.37981518049123e-06 91 0.091
+7.332478836389217e-06 92 0.092
+7.271541314832622e-06 93 0.093
+7.192482240725945e-06 94 0.094
+7.10747615998429e-06 95 0.095
+7.031089263809311e-06 96 0.096
+6.966464537344899e-06 97 0.097
+6.902700619786256e-06 98 0.098
+6.829339712022136e-06 99 0.099
+6.746047328463436e-06 100 0.1
