--- reg_tests/test_files/kovasznay_P7/kovasznay_P7.norm.gold
@@ -1,100 +1,0 @@
-8.56232727521058e-05 1 0.001
-0.0001057742942060063 2 0.002
-2.65363129570839e-05 3 0.003
-3.81522948081228e-06 4 0.004
-3.034750794880674e-06 5 0.005
-2.941116221384011e-06 6 0.006
-3.01386971325589e-06 7 0.007
-2.58206041529588e-06 8 0.008
-1.704167678382195e-06 9 0.009
-8.112513468835172e-07 10 0.01
-8.801666697779443e-07 11 0.011
-1.376724755016898e-06 12 0.012
-1.578878566399229e-06 13 0.013
-1.448872409431469e-06 14 0.014
-1.080975746257139e-06 15 0.015
-6.176009397632494e-07 16 0.016
-3.529655106728835e-07 17 0.017
-5.645320293864362e-07 18 0.018
-7.456066880581062e-07 19 0.019
-7.632938879850015e-07 20 0.02
-6.382181822178047e-07 21 0.021
-4.313361619652561e-07 22 0.022
-2.421062257738695e-07 23 0.023
-2.530132619553174e-07 24 0.024
-3.572064645343997e-07 25 0.025
-4.014751245628142e-07 26 0.026
-3.704400789732514e-07 27 0.027
-2.844399696233391e-07 28 0.028
-1.801726650218283e-07 29 0.029
-1.248054875272466e-07 30 0.03
-1.613538031127134e-07 31 0.031
-1.980027345056584e-07 32 0.032
-2.006136697551246e-07 33 0.033
-1.721327339418839e-07 34 0.034
-1.27543668752266e-07 35 0.035
-9.201641945653639e-08 36 0.036
-9.30857985329817e-08 37 0.037
-1.102293653749072e-07 38 0.038
-1.167954591318849e-07 39 0.039
-1.08033001049347e-07 40 0.04
-8.831268332954209e-08 41 0.041
-6.673449188778054e-08 42 0.042
-5.632198797879246e-08 43 0.043
-6.058888689829367e-08 44 0.044
-6.601536875127926e-08 45 0.045
-6.555394000965528e-08 46 0.046
-5.950558925459848e-08 47 0.047
-5.137991086951018e-08 48 0.048
-4.599018084445241e-08 49 0.049
-4.552159883309028e-08 50 0.05
-4.6862769743695e-08 51 0.051
-4.660528055524094e-08 52 0.052
-4.388205652811362e-08 53 0.053
-3.969439898617998e-08 54 0.054
-3.593920903152817e-08 55 0.055
-3.405360470941606e-08 56 0.056
-3.38289884363476e-08 57 0.057
-3.381802335321204e-08 58 0.058
-3.315663435858881e-08 59 0.059
-3.190169889655047e-08 60 0.06
-3.055655214781754e-08 61 0.061
-2.95734435044656e-08 62 0.062
-2.900214068659227e-08 63 0.063
-2.850435619562883e-08 64 0.064
-2.776866908598687e-08 65 0.065
-2.682075462224826e-08 66 0.066
-2.578101661524564e-08 67 0.067
-2.493995482440548e-08 68 0.068
-2.433388501787767e-08 69 0.069
-2.383954029933524e-08 70 0.07
-2.342045922981376e-08 71 0.071
-2.292424193069589e-08 72 0.072
-2.236407930663588e-08 73 0.073
-2.187034531638665e-08 74 0.074
-2.139578255197762e-08 75 0.075
-2.094151518640901e-08 76 0.076
-2.048002547269842e-08 77 0.077
-1.995456822571251e-08 78 0.078
-1.948806468938108e-08 79 0.079
-1.902694787994883e-08 80 0.08
-1.860999418675781e-08 81 0.081
-1.821095299844052e-08 82 0.082
-1.78463551717119e-08 83 0.083
-1.750219654891686e-08 84 0.084
-1.712536852691981e-08 85 0.085
-1.677272984892148e-08 86 0.086
-1.643618809613962e-08 87 0.087
-1.608620231379668e-08 88 0.088
-1.575829599116238e-08 89 0.089
-1.540768508164154e-08 90 0.09
-1.507391267605679e-08 91 0.091
-1.480307670022945e-08 92 0.092
-1.445631274276544e-08 93 0.093
-1.418909430252078e-08 94 0.094
-1.390696673235734e-08 95 0.095
-1.361548626025945e-08 96 0.096
-1.338930862893949e-08 97 0.097
-1.311783189705441e-08 98 0.098
-1.288330276802216e-08 99 0.099
-1.258034301549239e-08 100 0.1
