+++ src/master_element/Hex8FEM.C
@@ -8,8 +8,8 @@
 
 #include <master_element/Hex8FEM.h>
 #include <master_element/MasterElement.h>
-
 #include <master_element/MasterElementFunctions.h>
+#include <master_element/TensorOps.h>
 
 #include <FORTRAN_Proto.h>
 
@@ -18,6 +18,87 @@
 
 namespace sierra{
 namespace nalu{
+
+namespace {
+
+  template<typename AlgTraits, typename GradViewType,  typename CoordViewType,
+           typename DetjType, typename OutputViewType>
+  void generic_grad_op_3d(
+    GradViewType referenceGradWeights,
+    CoordViewType coords,
+    OutputViewType weights,
+    DetjType detj)
+  {
+    /**
+     * Given the reference gradient weights evaluated at the integration points and the cooordinates,
+     * this method computes
+     *     \nabla_\mathbf{x}|_{\alpha}  as \left( J^{-T} \nabla_{\mathbf{x}^\hat} \right)|_\alpha,
+     *
+     *  This operation can be specialized for efficiency on hex topologies (as tensor-contractions)
+     *  or on tets (since the gradient is independent of \alpha).  But this can work as a fallback.
+     *
+     *  Also saves detj at the integration points---useful for FEM but not used at all in CVFEM
+     */
+
+    using ftype = typename CoordViewType::value_type;
+    static_assert(std::is_same<ftype, typename GradViewType::value_type>::value, "Incompatiable value type for views");
+    static_assert(std::is_same<ftype, typename OutputViewType::value_type>::value, "Incompatiable value type for views");
+    static_assert(GradViewType::Rank == 3, "grad view assumed to be 3D");
+    static_assert(CoordViewType::Rank == 2, "Coordinate view assumed to be 2D");
+    static_assert(OutputViewType::Rank == 3, "Weight view assumed to be 3D");
+    static_assert(AlgTraits::nDim_ == 3, "3D method");
+
+    ThrowAssert(AlgTraits::nodesPerElement_ == referenceGradWeights.extent(1));
+    ThrowAssert(AlgTraits::nDim_ == referenceGradWeights.extent(2));
+    ThrowAssert(weights.extent(0) == referenceGradWeights.extent(0));
+    ThrowAssert(weights.extent(1) == referenceGradWeights.extent(1));
+    ThrowAssert(weights.extent(2) == referenceGradWeights.extent(2));
+
+    for (unsigned ip = 0; ip < AlgTraits::numGp_; ++ip) {
+      ftype jact[3][3] = { {0.0, 0.0, 0.0}, {0.0, 0.0, 0.0}, {0.0, 0.0, 0.0} };
+
+      // compute Jacobian.  Stash away a local copy of the reference weights
+      // since they're be used again soon
+      ftype refGrad[AlgTraits::nodesPerElement_][3];
+      for (int n = 0; n < AlgTraits::nodesPerElement_; ++n) {
+        refGrad[n][0] = referenceGradWeights(ip, n, 0);
+        refGrad[n][1] = referenceGradWeights(ip, n, 1);
+        refGrad[n][2] = referenceGradWeights(ip, n, 2);
+
+        jact[0][0] += refGrad[n][0] * coords(n, 0);
+        jact[0][1] += refGrad[n][1] * coords(n, 0);
+        jact[0][2] += refGrad[n][2] * coords(n, 0);
+
+        jact[1][0] += refGrad[n][0] * coords(n, 1);
+        jact[1][1] += refGrad[n][1] * coords(n, 1);
+        jact[1][2] += refGrad[n][2] * coords(n, 1);
+
+        jact[2][0] += refGrad[n][0] * coords(n, 2);
+        jact[2][1] += refGrad[n][1] * coords(n, 2);
+        jact[2][2] += refGrad[n][2] * coords(n, 2);
+      }
+
+      ftype invJac[3][3];
+      adjugate_matrix33(jact, invJac);
+
+      detj(ip) = jact[0][0] * invJac[0][0] + jact[1][0] * invJac[1][0] + jact[2][0] * invJac[2][0];
+      ThrowAssertMsg(stk::simd::are_any(detj(ip) > +tiny_positive_value()),"Problem with determinant");
+
+      const ftype inv_detj = ftype(1.0) / detj(ip);
+      for (int d_outer = 0; d_outer < 3; ++d_outer) {
+        for (int d_inner = 0; d_inner < 3; ++d_inner) {
+          invJac[d_outer][d_inner] *= inv_detj;
+        }
+      }
+
+      for (int n = 0; n < AlgTraits::nodesPerElement_; ++n) {
+        weights(ip, n, 0) = invJac[0][0] * refGrad[n][0] + invJac[0][1] * refGrad[n][1] + invJac[0][2] * refGrad[n][2];
+        weights(ip, n, 1) = invJac[1][0] * refGrad[n][0] + invJac[1][1] * refGrad[n][1] + invJac[1][2] * refGrad[n][2];
+        weights(ip, n, 2) = invJac[2][0] * refGrad[n][0] + invJac[2][1] * refGrad[n][1] + invJac[2][2] * refGrad[n][2];
+      }
+    }
+  }
+}
 
 //--------------------------------------------------------------------------
 //-------- constructor -----------------------------------------------------
@@ -99,17 +180,14 @@ void Hex8FEM::grad_op(
 //--------------------------------------------------------------------------
 //-------- grad_op ---------------------------------------------------------
 //--------------------------------------------------------------------------
-void Hex8FEM::grad_op(
+void Hex8FEM::grad_op_fem(
   SharedMemView<DoubleType**>&coords,
   SharedMemView<DoubleType***>&gradop,
   SharedMemView<DoubleType***>&deriv,
-  SharedMemView<DoubleType*>&det_j,
-  DoubleType &error)
+  SharedMemView<DoubleType*>&det_j)
 {
   hex8_fem_derivative(numIntPoints_, &intgLoc_[0], deriv);
   generic_grad_op_3d<AlgTraitsHex8>(deriv, coords, gradop, det_j);
-
-  error = 0.0; // error checking in generic_grad_op
 }
 
 //--------------------------------------------------------------------------
@@ -139,17 +217,14 @@ void Hex8FEM::shifted_grad_op(
 //--------------------------------------------------------------------------
 //-------- shifted_grad_op -------------------------------------------------
 //--------------------------------------------------------------------------
-void Hex8FEM::shifted_grad_op(
+void Hex8FEM::shifted_grad_op_fem(
   SharedMemView<DoubleType**>&coords,
   SharedMemView<DoubleType***>&gradop,
   SharedMemView<DoubleType***>&deriv,
-  SharedMemView<DoubleType*>&det_j,
-  DoubleType &error)
+  SharedMemView<DoubleType*>&det_j)
 {
   hex8_fem_derivative(numIntPoints_, &intgLocShift_[0], deriv);
   generic_grad_op_3d<AlgTraitsHex8>(deriv, coords, gradop, det_j);
-
-  error = 0.0; // error checking in generic_grad_op
 }
 
 //--------------------------------------------------------------------------
@@ -204,8 +279,8 @@ void
 Hex8FEM::general_shape_fcn(
   const int numIp,
   const double *isoParCoord,
-  double *shpfc) {
-
+  double *shpfc)
+{
   hex8_fem_shape_fcn(numIp,isoParCoord,shpfc);
 }
 
@@ -330,7 +405,7 @@ Hex8FEM::hex8_fem_derivative(
     deriv[i*nodesPerElement_*3 + 3] =  0.125*(1.0-par_coord[i*3+1])*(1.0-par_coord[i*3+2]);
     deriv[i*nodesPerElement_*3 + 6] =  0.125*(1.0+par_coord[i*3+1])*(1.0-par_coord[i*3+2]);
     deriv[i*nodesPerElement_*3 + 9] = -0.125*(1.0+par_coord[i*3+1])*(1.0-par_coord[i*3+2]);
-    deriv[i*nodesPerElement_*3 + 12  ] = -0.125*(1.0-par_coord[i*3+1])*(1.0+par_coord[i*3+2]);
+    deriv[i*nodesPerElement_*3 + 12] = -0.125*(1.0-par_coord[i*3+1])*(1.0+par_coord[i*3+2]);
     deriv[i*nodesPerElement_*3 + 15] =  0.125*(1.0-par_coord[i*3+1])*(1.0+par_coord[i*3+2]);
     deriv[i*nodesPerElement_*3 + 18] =  0.125*(1.0+par_coord[i*3+1])*(1.0+par_coord[i*3+2]);
     deriv[i*nodesPerElement_*3 + 21] = -0.125*(1.0+par_coord[i*3+1])*(1.0+par_coord[i*3+2]);
@@ -392,7 +467,6 @@ Hex8FEM::hex8_fem_derivative(
     deriv(ip,6,2) =  0.125*(1.0+x)*(1.0+y);
     deriv(ip,7,2) =  0.125*(1.0-x)*(1.0+y);
   }
-
 }
 
 }
