+++ src/master_element/MasterElementFactory.C
@@ -1,0 +1,250 @@
+/*------------------------------------------------------------------------*/
+/*  Copyright 2014 Sandia Corporation.                                    */
+/*  This software is released under the license detailed                  */
+/*  in the file, LICENSE, which is located in the top-level Nalu          */
+/*  directory structure                                                   */
+/*------------------------------------------------------------------------*/
+
+#include <master_element/MasterElementFactory.h>
+#include <master_element/MasterElement.h>
+
+#include <master_element/Hex8CVFEM.h>
+#include <master_element/Hex27CVFEM.h>
+#include <master_element/MasterElementHO.h>
+
+#include <NaluEnv.h>
+#include <nalu_make_unique.h>
+
+#include <stk_util/environment/ReportHandler.hpp>
+#include <stk_topology/topology.hpp>
+
+#include <cmath>
+#include <iostream>
+#include <memory>
+
+namespace sierra{
+namespace nalu{
+
+  std::unique_ptr<MasterElement>
+  create_surface_master_element(stk::topology topo)
+  {
+    if (topo.is_super_topology()) {
+      // super topologies uses different master element type
+      return nullptr;
+    }
+
+    switch ( topo.value() ) {
+
+      case stk::topology::HEX_8:
+        return make_unique<HexSCS>();
+
+      case stk::topology::HEX_27:
+        return make_unique<Hex27SCS>();
+
+      case stk::topology::TET_4:
+        return make_unique<TetSCS>();
+
+      case stk::topology::PYRAMID_5:
+        return make_unique<PyrSCS>();
+
+      case stk::topology::WEDGE_6:
+        return make_unique<WedSCS>();
+
+      case stk::topology::QUAD_4:
+        return make_unique<Quad3DSCS>();
+
+      case stk::topology::QUAD_9:
+        return make_unique<Quad93DSCS>();
+
+      case stk::topology::TRI_3:
+        return make_unique<Tri3DSCS>();
+
+      case stk::topology::QUAD_4_2D:
+        return make_unique<Quad2DSCS>();
+
+      case stk::topology::QUAD_9_2D:
+        return make_unique<Quad92DSCS>();
+
+      case stk::topology::TRI_3_2D:
+        return make_unique<Tri2DSCS>();
+
+      case stk::topology::LINE_2:
+        return make_unique<Edge2DSCS>();
+
+      case stk::topology::LINE_3:
+        return make_unique<Edge32DSCS>();
+
+      case stk::topology::SHELL_QUAD_4:
+        NaluEnv::self().naluOutputP0() << "SHELL_QUAD_4 only supported for io surface transfer applications" << std::endl;
+        return make_unique<Quad3DSCS>();
+
+      case stk::topology::SHELL_TRI_3:
+        NaluEnv::self().naluOutputP0() << "SHELL_TRI_3 only supported for io surface transfer applications" << std::endl;
+        return make_unique<Tri3DSCS>();
+
+      default:
+        NaluEnv::self().naluOutputP0() << "sorry, we only support hex8, tet4, pyr5, wed6,"
+                                          " quad2d, quad3d, tri2d, tri3d and edge2d surface elements" << std::endl;
+        NaluEnv::self().naluOutputP0() << "your type is " << topo.value() << std::endl;
+        break;
+
+    }
+    return nullptr;
+  }
+  //--------------------------------------------------------------------------
+  std::unique_ptr<MasterElement>
+  create_volume_master_element(stk::topology topo)
+  {
+    if (topo.is_super_topology()) {
+      // super topologies uses different master element type
+      return nullptr;
+    }
+
+    switch ( topo.value() ) {
+
+      case stk::topology::HEX_8:
+        return make_unique<HexSCV>();
+
+      case stk::topology::HEX_27:
+        return make_unique<Hex27SCV>();
+
+      case stk::topology::TET_4:
+        return make_unique<TetSCV>();
+
+      case stk::topology::PYRAMID_5:
+        return make_unique<PyrSCV>();
+
+      case stk::topology::WEDGE_6:
+        return  make_unique<WedSCV>();
+
+      case stk::topology::QUAD_4_2D:
+        return make_unique<Quad2DSCV>();
+
+      case stk::topology::QUAD_9_2D:
+        return make_unique<Quad92DSCV>();
+
+      case stk::topology::TRI_3_2D:
+        return make_unique<Tri2DSCV>();
+
+      default:
+        NaluEnv::self().naluOutputP0() << "sorry, we only support hex8, tet4, wed6, "
+                                          " pyr5, quad4, and tri3 volume elements" << std::endl;
+        NaluEnv::self().naluOutputP0() << "your type is " << topo.value() << std::endl;
+        break;
+    }
+    return nullptr;
+  }
+  //--------------------------------------------------------------------------
+  std::unique_ptr<MasterElement>
+  create_surface_master_element(stk::topology topo, int dimension, std::string quadType)
+  {
+    if (!topo.is_super_topology()) {
+      // regular topologies uses different master element type
+      return nullptr;
+    }
+
+    ThrowRequire(dimension > 0 && dimension < 4);
+
+    auto desc = ElementDescription::create(dimension, topo);
+
+    auto basis = (topo.is_superelement()) ?
+        LagrangeBasis(desc->inverseNodeMap, desc->nodeLocs1D)
+      : LagrangeBasis(desc->inverseNodeMapBC, desc->nodeLocs1D);
+
+    auto quad = TensorProductQuadratureRule(quadType, desc->polyOrder);
+
+    if (topo.is_superedge()) {
+      ThrowRequire(desc->baseTopo == stk::topology::QUAD_4_2D);
+      return make_unique<HigherOrderEdge2DSCS>(*desc, basis, quad);
+    }
+
+    if (topo.is_superface()) {
+      ThrowRequire(desc->baseTopo == stk::topology::HEX_8);
+      return make_unique<HigherOrderQuad3DSCS>(*desc, basis, quad);
+    }
+
+    if (topo.is_superelement() && desc->baseTopo == stk::topology::QUAD_4_2D) {
+      return make_unique<HigherOrderQuad2DSCS>(*desc, basis, quad);
+    }
+
+    if (topo.is_superelement() && desc->baseTopo == stk::topology::HEX_8) {
+      return make_unique<HigherOrderHexSCS>(*desc, basis, quad);
+    }
+
+    return nullptr;
+  }
+  //--------------------------------------------------------------------------
+  std::unique_ptr<MasterElement>
+  create_volume_master_element(
+    stk::topology topo,
+    int dimension,
+    std::string quadType)
+  {
+    if (!topo.is_super_topology()) {
+      // regular topologies uses different master element type
+      return nullptr;
+    }
+
+    ThrowRequire(dimension > 0 && dimension < 4);
+
+    auto desc = ElementDescription::create(dimension, topo);
+    auto basis = LagrangeBasis(desc->inverseNodeMap, desc->nodeLocs1D);
+    auto quad = TensorProductQuadratureRule(quadType, desc->polyOrder);
+
+    switch (desc->baseTopo.value()) {
+      case stk::topology::QUADRILATERAL_4_2D:
+        return make_unique<HigherOrderQuad2DSCV>(*desc, basis, quad);
+      case stk::topology::HEXAHEDRON_8:
+        return make_unique<HigherOrderHexSCV>(*desc, basis, quad);
+      default:
+        NaluEnv::self().naluOutputP0() << "High order elements only support base quad4 and hex8 meshes" << std::endl;
+        break;
+    }
+    return nullptr;
+  }
+
+
+  MasterElement*
+  get_surface_master_element(
+    const stk::topology& theTopo,
+    int dimension,
+    std::string quadType)
+  {
+    static std::map<stk::topology, std::unique_ptr<MasterElement>> surfaceMeMap;
+
+    auto it = surfaceMeMap.find(theTopo);
+    if (it == surfaceMeMap.end()) {
+      if (!theTopo.is_super_topology()) {
+        surfaceMeMap[theTopo] = create_surface_master_element(theTopo);
+      } else {
+        surfaceMeMap[theTopo] = create_surface_master_element(theTopo, dimension, quadType);
+      }
+    }
+    MasterElement* theElem = surfaceMeMap.at(theTopo).get();
+    ThrowRequire(theElem != nullptr);
+    return theElem;
+  }
+
+  MasterElement*
+  get_volume_master_element(
+    const stk::topology& theTopo,
+    int dimension,
+    std::string quadType)
+  {
+    static std::map<stk::topology, std::unique_ptr<MasterElement>> volumeMeMap;
+
+    auto it = volumeMeMap.find(theTopo);
+    if (it == volumeMeMap.end()) {
+      if (!theTopo.is_super_topology()) {
+        volumeMeMap[theTopo] = create_volume_master_element(theTopo);
+      } else {
+        volumeMeMap[theTopo] = create_volume_master_element(theTopo, dimension, quadType);
+      }
+    }
+    MasterElement* theElem = volumeMeMap.at(theTopo).get();
+    ThrowRequire(theElem != nullptr);
+    return theElem;
+  }
+
+}
+}
