+++ unit_tests/UnitTestHOMasterElements.C
@@ -12,7 +12,7 @@
 #include <stk_mesh/base/FieldBase.hpp>
 
 #include <element_promotion/ElementDescription.h>
-#include <element_promotion/MasterElementHO.h>
+#include <master_element/MasterElementHO.h>
 #include <element_promotion/QuadratureRule.h>
 #include <element_promotion/LagrangeBasis.h>
 #include <element_promotion/TensorProductQuadratureRule.h>
